+++ lib/main.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:sikcal/components/loading_widget.dart';
 
 import 'package:sikcal/screen/welcome/welcome_screen.dart';
 import 'package:sikcal/data/constants.dart';
-import 'package:sikcal/screens/loading_screen.dart';
 import 'package:sikcal/screens/main_view.dart';
 
 void main() {
