+++ lib/screens/components/meal_list_view.dart
@@ -4,7 +4,7 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:sikcal/data/constants.dart';
 import 'package:sikcal/data/provider.dart';
 import 'package:sikcal/model/meal.dart';
-import 'package:sikcal/screens/search_menu_view.dart';
+import 'package:sikcal/screens/home/search_menu_view.dart';
 
 class MealListView extends ConsumerWidget {
   List<Meal> mealList;
