+++ lib/screens/main_view.dart
@@ -4,8 +4,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:sikcal/data/constants.dart';
 import 'package:sikcal/screen/mypage/mypage_main_view.dart';
-import 'package:sikcal/screens/home_view.dart';
 import 'package:sikcal/screen/mydiet/mydiet_main_view.dart';
+import 'package:sikcal/screens/home/home_view.dart';
 
 class MainView extends ConsumerStatefulWidget {
   const MainView({Key? key}) : super(key: key);
